fix : jdbc template 갖다버림, jpa 적용, 내친김에 spring data jpa도 적용, jwt 보일러 플레이트 작업

+++ src/main/java/one/sentence_spring/todo/dto/AddTodoRequest.java
@@ -1,4 +1,4 @@
 package one.sentence_spring.todo.dto;
 
-public record AddTodoRequest(String content, int folderId, int userId, int index) {
+public record AddTodoRequest(String content, Long folderId, Long userId, int index) {
 }
